Refactor editor rendering
